arranco con pedir turno

+++ src/Clinica Frba/Clases/Turno.cs
@@ -7,11 +7,11 @@ namespace Clinica_Frba.Clases
 {
     class Turno
     {
-        public int Id { get; set; }
-        public int Codigo_Persona { get; set; }
-        public int Codigo_Profesional { get; set; }
+        public decimal Id { get; set; }
+        public decimal Codigo_Persona { get; set; }
+        public decimal Codigo_Profesional { get; set; }
         public DateTime Fecha { get; set; }
-        public int Codigo_Especialidad { get; set; }
+        public decimal Codigo_Especialidad { get; set; }
         public bool Estado { get; set; }
         public DateTime Horario_Llegada { get; set; }
     }
